Promote RefCountedReleasable to top level
